Graphing Regression
The regression line is now graphed. It is a known bug that the regression line can travel outside of the chart, but in the context of my personal data, it will probably never do that.

If I want to make this into a full-on Chart js plugin, I'll have to solve that big problem. But horray, because I solved a big issue I had with the charts!

+++ src/pages/personal/chart.ts
@@ -59,7 +59,7 @@ export const GET: APIRoute = async ({ url }) => {
 
         // save the regression module's life by parsing every integer. IDK why this boi hates me for inputting strings.
         unprocessedContent.map((value: string[], index: number) => {
-            content.push([parseInt(value[0]), parseInt(value[1])]);
+            content.push([parseFloat(value[0]), parseFloat(value[1])]);
         })
 
         // sort by X-axis value
